submissions/leopoldo-lima/solution/requirements-audit.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pytest>=8.0.0
pytest-httpx>=0.35.0
ruff>=0.6.0
mypy>=1.10.0
pip-audit>=2.7.0
pre-commit>=3.8.0
cyclonedx-bom>=4.0.0
